// server/vance-addon-brain-issues/client/src/generated/issues/IssuesView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.issues.IssuesView
*/

import { IssueView } from './IssueView';
import { IssueStatsView } from './IssueStatsView';

export interface IssuesView {
  folder: string;
  title?: string;
  description?: string;
  labels: string[];
  issues: IssueView[];
  stats: IssueStatsView;
  archivedView: boolean;
}