// server/vance-addon-brain-issues/client/src/generated/issues/IssueContentView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.issues.IssueContentView
*/

import { IssueCommentView } from './IssueCommentView';

export interface IssueContentView {
  id: string;
  path: string;
  number: number;
  title: string;
  state: string;
  labels: string[];
  assignee?: string;
  priority?: string;
  archived: boolean;
  body: string;
  comments: IssueCommentView[];
}